+++ src/main/java/com/giassi/microservice/demo2/rest/GlobalExceptionHandler.java
@@ -13,7 +13,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler({InvalidEmailException.class, InvalidGenderException.class, InvalidUserDataException.class,
             InvalidUserIdentifierException.class, InvalidRoleIdentifierException.class, InvalidUsernameException.class,
             InvalidLoginException.class, InvalidPermissionDataException.class, InvalidRoleDataException.class,
-            RoleInUseException.class})
+            RoleInUseException.class, PermissionInUseException.class})
     public ResponseEntity<ErrorDetails> handleAsBadRequest(RuntimeException ex) {
         ErrorDetails errorDetails = new ErrorDetails(ex.getMessage());
         return new ResponseEntity<>(errorDetails, HttpStatus.BAD_REQUEST);