+++ src/main/java/com/giassi/microservice/demo2/rest/users/dtos/PermissionDTO.java
@@ -10,10 +10,14 @@ public class PermissionDTO implements java.io.Serializable {
 
     private Long id;
     private String permission;
+    private boolean enabled;
+    private String note;
 
     public PermissionDTO(Permission permission) {
         this.id = permission.getId();
         this.permission = permission.getPermission();
+        this.enabled = permission.isEnabled();
+        this.note = permission.getNote();
     }
 
 }