+++ src/main/java/com/giassi/microservice/demo2/rest/users/exceptions/PermissionInUseException.java
@@ -1,0 +1,9 @@
+package com.giassi.microservice.demo2.rest.users.exceptions;
+
+public class PermissionInUseException extends RuntimeException {
+
+    public PermissionInUseException(String message) {
+        super(message);
+    }
+
+}